+++ src/index.ts
@@ -1,4 +1,3 @@
-export * from './adapters/groupmq-bullboard-adapter';
 export * from './helpers';
 export * from './job';
 export * from './queue';